Factored out ProjectDetailsCreator
